+++ pillars/tenant/src/dto/update-tenant-group.dto.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/swagger';
+import { CreateTenantGroupDto } from './create-tenant-group.dto';
+
+export class UpdateTenantGroupDto extends PartialType(CreateTenantGroupDto) {}